[core] XPathHandler/XPathFunctionDefintion: Refactoring, javadoc
