+++ pmd-core/src/main/java/net/sourceforge/pmd/lang/rule/xpath/internal/SaxonXPathRuleQuery.java
@@ -5,14 +5,11 @@
 package net.sourceforge.pmd.lang.rule.xpath.internal;
 
 import java.util.ArrayList;
-import java.util.Arrays;
 import java.util.HashMap;
 import java.util.LinkedHashSet;
 import java.util.List;
 import java.util.Map;
 import java.util.Set;
-import java.util.stream.Collectors;
-import javax.xml.namespace.QName;
 
 import org.apache.commons.lang3.exception.ContextedRuntimeException;
 import org.slf4j.Logger;
@@ -24,7 +21,6 @@ import net.sourceforge.pmd.lang.rule.xpath.PmdXPathException;
 import net.sourceforge.pmd.lang.rule.xpath.PmdXPathException.Phase;
 import net.sourceforge.pmd.lang.rule.xpath.XPathVersion;
 import net.sourceforge.pmd.lang.rule.xpath.impl.XPathFunctionDefinition;
-import net.sourceforge.pmd.lang.rule.xpath.impl.XPathFunctionException;
 import net.sourceforge.pmd.lang.rule.xpath.impl.XPathHandler;
 import net.sourceforge.pmd.properties.PropertyDescriptor;
 import net.sourceforge.pmd.util.DataMap;
@@ -33,33 +29,19 @@ import net.sourceforge.pmd.util.DataMap.SimpleDataKey;
 import net.sf.saxon.Configuration;
 import net.sf.saxon.expr.Expression;
 import net.sf.saxon.expr.LocalVariableReference;
-import net.sf.saxon.expr.StaticContext;
-import net.sf.saxon.expr.StringLiteral;
-import net.sf.saxon.expr.XPathContext;
-import net.sf.saxon.lib.ExtensionFunctionCall;
 import net.sf.saxon.lib.ExtensionFunctionDefinition;
 import net.sf.saxon.lib.NamespaceConstant;
 import net.sf.saxon.om.AtomicSequence;
-import net.sf.saxon.om.EmptyAtomicSequence;
 import net.sf.saxon.om.Item;
 import net.sf.saxon.om.NamePool;
-import net.sf.saxon.om.Sequence;
 import net.sf.saxon.om.SequenceIterator;
 import net.sf.saxon.om.StructuredQName;
-import net.sf.saxon.pattern.NodeKindTest;
 import net.sf.saxon.sxpath.IndependentContext;
 import net.sf.saxon.sxpath.XPathDynamicContext;
 import net.sf.saxon.sxpath.XPathEvaluator;
 import net.sf.saxon.sxpath.XPathExpression;
 import net.sf.saxon.sxpath.XPathVariable;
 import net.sf.saxon.trans.XPathException;
-import net.sf.saxon.value.BigDecimalValue;
-import net.sf.saxon.value.BooleanValue;
-import net.sf.saxon.value.EmptySequence;
-import net.sf.saxon.value.Int64Value;
-import net.sf.saxon.value.SequenceExtent;
-import net.sf.saxon.value.SequenceType;
-import net.sf.saxon.value.StringValue;
 
 
 /**
@@ -218,7 +200,7 @@ public class SaxonXPathRuleQuery {
         }
 
         for (XPathFunctionDefinition xpathFun : xPathHandler.getRegisteredExtensionFunctions()) {
-            ExtensionFunctionDefinition fun = convertAbstractXPathFunctionDefinition(xpathFun);
+            ExtensionFunctionDefinition fun = new SaxonExtensionFunctionDefinitionAdapter(xpathFun);
             StructuredQName qname = fun.getFunctionQName();
             staticCtx.declareNamespace(qname.getPrefix(), qname.getURI());
             this.configuration.registerExtensionFunction(fun);
@@ -302,133 +284,5 @@ public class SaxonXPathRuleQuery {
             local.setStaticType(null, converted, 0);
             return local;
         }
-    }
-
-    public static ExtensionFunctionDefinition convertAbstractXPathFunctionDefinition(XPathFunctionDefinition definition) {
-        final SequenceType SINGLE_ELEMENT_SEQUENCE_TYPE = NodeKindTest.ELEMENT.one();
-
-        return new ExtensionFunctionDefinition() {
-            private SequenceType convertToSequenceType(XPathFunctionDefinition.Type type) {
-                switch (type) {
-                    case SINGLE_STRING: return SequenceType.SINGLE_STRING;
-                    case SINGLE_BOOLEAN: return SequenceType.SINGLE_BOOLEAN;
-                    case SINGLE_ELEMENT: return SINGLE_ELEMENT_SEQUENCE_TYPE;
-                    case SINGLE_INTEGER: return SequenceType.SINGLE_INTEGER;
-                    case STRING_SEQUENCE: return SequenceType.STRING_SEQUENCE;
-                    case OPTIONAL_STRING: return SequenceType.OPTIONAL_STRING;
-                    case OPTIONAL_DECIMAL: return SequenceType.OPTIONAL_DECIMAL;
-                    default:
-                        throw new UnsupportedOperationException("Type " + type + " is not supported");
-                }
-            }
-
-            private SequenceType[] convertToSequenceTypes(XPathFunctionDefinition.Type[] types) {
-                SequenceType[] result = new SequenceType[types.length];
-                for (int i = 0; i < types.length; i++) {
-                    result[i] = convertToSequenceType(types[i]);
-                }
-                return result;
-            }
-
-            @Override
-            public StructuredQName getFunctionQName() {
-                QName qName = definition.getQName();
-                return new StructuredQName(qName.getPrefix(), qName.getNamespaceURI(), qName.getLocalPart());
-            }
-
-            @Override
-            public SequenceType[] getArgumentTypes() {
-                return convertToSequenceTypes(definition.getArgumentTypes());
-            }
-
-            @Override
-            public SequenceType getResultType(SequenceType[] suppliedArgumentTypes) {
-                return convertToSequenceType(definition.getResultType());
-            }
-
-            @Override
-            public boolean dependsOnFocus() {
-                return definition.dependsOnContext();
-            }
-
-            @Override
-            public ExtensionFunctionCall makeCallExpression() {
-                XPathFunctionDefinition.FunctionCall call = definition.makeCallExpression();
-                return new ExtensionFunctionCall() {
-                    @Override
-                    public Expression rewrite(StaticContext context, Expression[] arguments) throws XPathException {
-                        Object[] convertedArguments = new Object[definition.getArgumentTypes().length];
-                        for (int i = 0; i < convertedArguments.length; i++) {
-                            if (arguments[i] instanceof StringLiteral) {
-                                convertedArguments[i] = ((StringLiteral) arguments[i]).getStringValue();
-                            }
-                        }
-                        try {
-                            call.staticInit(convertedArguments);
-                        } catch (XPathFunctionException e) {
-                            XPathException xPathException = new XPathException(e);
-                            xPathException.setIsStaticError(true);
-                            throw xPathException;
-                        }
-                        return null;
-                    }
-
-                    @Override
-                    public Sequence call(XPathContext context, Sequence[] arguments) throws XPathException {
-                        Node contextNode = null;
-                        if (definition.dependsOnContext()) {
-                            contextNode = XPathElementToNodeHelper.itemToNode(context.getContextItem());
-                        }
-                        Object[] convertedArguments = new Object[definition.getArgumentTypes().length];
-                        for (int i = 0; i < convertedArguments.length; i++) {
-                            switch (definition.getArgumentTypes()[i]) {
-                                case SINGLE_STRING:
-                                    convertedArguments[i] = arguments[i].head().getStringValue();
-                                    break;
-                                case SINGLE_ELEMENT:
-                                    convertedArguments[i] = arguments[i].head();
-                                    break;
-                                default:
-                                throw new UnsupportedOperationException("Don't know how to convert argument type " + definition.getArgumentTypes()[i]);
-                            }
-                        }
-
-
-                        Object result = null;
-                        try {
-                            result = call.call(contextNode, convertedArguments);
-                        } catch (XPathFunctionException e) {
-                            throw new XPathException(e);
-                        }
-                        Sequence convertedResult = null;
-                        switch (definition.getResultType()) {
-                            case SINGLE_BOOLEAN:
-                                convertedResult = BooleanValue.get((Boolean) result);
-                                break;
-                            case SINGLE_INTEGER:
-                                convertedResult = Int64Value.makeIntegerValue((Integer) result);
-                                break;
-                            case SINGLE_STRING:
-                                convertedResult = new StringValue((String) result);
-                                break;
-                            case OPTIONAL_STRING:
-                                convertedResult = result != null ? new StringValue((String) result) : EmptyAtomicSequence.INSTANCE;
-                                break;
-                            case STRING_SEQUENCE:
-                                convertedResult = result != null ? new SequenceExtent(Arrays.stream((String[]) result).map(StringValue::new).collect(Collectors.toList()))
-                                        : EmptySequence.getInstance();
-                                break;
-                            case OPTIONAL_DECIMAL:
-                                convertedResult = result != null && Double.isFinite((Double) result) ? new BigDecimalValue((Double) result)
-                                        : EmptySequence.getInstance();
-                                break;
-                            default:
-                                throw new UnsupportedOperationException("Don't know how to convert result type " + definition.getResultType());
-                        }
-                        return convertedResult;
-                    }
-                };
-            }
-        };
     }
 }
